Creation of binomial generalised linear models with varying axes
I created 3 main graphs with varying axis to visualise the influence of temperature and exposure time on the survival rate of beetles. I attempted to work on the new graph, with probability projections but largely struggled, hopefully I can attempt this and come back to it in a new day. Some of the suggestions from how to atart this are included in the "Dan Suggestions" R studio file. The unfiltered graph is also first 'toy' around with the data before being introduced to the binomial figures.